some advances on assembly coding but no actual gain - yet

+++ avr-gcc/runaes.c
@@ -39,8 +39,8 @@ int main()
   for(i=0;i<16;i++) key[i]=pgm_read_byte(&(keyex[i]));
   for(i=0;i<16;i++) in[i]=pgm_read_byte(&(inex[i]));
 
-  double dt,base;
-
+  uint8_t dt,base;
+/*
   printf("Without countermeasure, plain: ");
   dt=run_aes(in,out,key,nt);
   base=dt;
@@ -52,17 +52,17 @@ int main()
   runalgo(&aes_rp,in,out,key,outex,16,nt,base);
   printMesSRAM("out",out);
 
-  
+  */
   for(n=3;n<=3;n+=2)
   {
     printf("n=%d\n",n);
-    
+    /*
     printf("  With RP countermeasure: ");
     dt=run_aes_share(in,out,key,n,&subbyte_rp_share,nt);
     report_time(dt,nt,base);
     check_ciphertext(out,outex,16);
     printMesSRAM("out",out);
-
+    */
     printf("  With randomized table : ");
     dt=run_aes_share(in,out,key,n,&subbyte_htable,nt);
     report_time(dt,nt,base);
